Extract editor file tree UI
