+++ assets/js/editor-main.js
@@ -1,6 +1,6 @@
 import { configureFetchCachePolicy } from './cache-control.js';
-import { createMarkdownBlocksEditor } from './editor-blocks.js?v=press-system-v3.4.24';
-import { createHiEditor } from './hieditor.js?v=press-system-v3.4.24';
+import { createMarkdownBlocksEditor } from './editor-blocks.js?v=press-system-v3.4.25';
+import { createHiEditor } from './hieditor.js?v=press-system-v3.4.25';
 import { insertImageMarkdownAtSelection, normalizeDateInputValue } from './editor-markdown-ops.js';
 import {
   FRONT_MATTER_FIELD_DEFS,
@@ -10,12 +10,12 @@ import {
   parseMarkdownFrontMatter,
   resolveFrontMatterBindings,
   valueIsPresent
-} from './frontmatter-document.js?v=press-system-v3.4.24';
-import { getContentRoot, resolveImageSrc } from './safe-html.js?v=press-system-v3.4.24';
-import { hydrateInternalLinkCards } from './link-cards.js?v=press-system-v3.4.24';
+} from './frontmatter-document.js?v=press-system-v3.4.25';
+import { getContentRoot, resolveImageSrc } from './safe-html.js?v=press-system-v3.4.25';
+import { hydrateInternalLinkCards } from './link-cards.js?v=press-system-v3.4.25';
 import { fetchConfigWithYamlFallback, fetchMergedSiteConfig } from './yaml.js';
-import { t, withLangParam, loadContentJsonWithRaw, getCurrentLang, normalizeLangKey } from './i18n.js?v=press-system-v3.4.24';
-import { resolveLocalMarkdownAssetReference } from './repository-deletions.js?v=press-system-v3.4.24';
+import { t, withLangParam, loadContentJsonWithRaw, getCurrentLang, normalizeLangKey } from './i18n.js?v=press-system-v3.4.25';
+import { resolveLocalMarkdownAssetReference } from './repository-deletions.js?v=press-system-v3.4.25';
 
 const LS_WRAP_KEY = 'press_editor_wrap_enabled';
 const LS_VIEW_KEY = 'press_editor_markdown_view_v2';
