+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.24",
-  "tag": "v3.4.24",
+  "version": "3.4.25",
+  "tag": "v3.4.25",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.23 <3.4.24"
+      ">=3.4.24 <3.4.25"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.23 first."
+    "message": "Update to v3.4.24 first."
   }
 }
