+++ index.html
@@ -69,11 +69,11 @@
       }
     }
   </style>
-  <script src="assets/js/theme-boot.js?v=press-system-v3.4.24"></script>
+  <script src="assets/js/theme-boot.js?v=press-system-v3.4.25"></script>
   <link rel="stylesheet" id="theme-pack">
 </head>
 <body>
   <div id="press-boot-progress" aria-hidden="true"><span></span></div>
-  <script type="module" src="assets/main.js?v=press-system-v3.4.24"></script>
+  <script type="module" src="assets/main.js?v=press-system-v3.4.25"></script>
 </body>
 </html>
